requirements.txt
boto3
PyJWT
stability-sdk[sagemaker] @ git+https://github.com/Stability-AI/stability-sdk.git@sagemaker
anthropic==0.2.10
streamlit-chat
streamlit-scrollable-textbox
streamlit_drawable_canvas
streamlit-cognito-auth
pydot
diagrams
gremlinpython
ai21==1.1.4
matplotlib
streamlit
sagemaker
textract
pandas
pillow
numpy
pypdf
pdf2image
bokeh==2.4.3
streamlit-bokeh-events
protobuf==3.20.*
streamlit-reveal-slides
streamlit-analytics
qrcode
python-jose
